requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
email-validator==1.3.1
Werkzeug==2.2.3
pywebview==4.3.3
pyinstaller==6.3.0
pillow==10.2.0
